requirements.txt
absl-py==1.0.0
aiohttp==3.8.3
aiosignal==1.3.1
argcomplete==2.0.0
async-timeout==4.0.2
asynctest==0.13.0
attributee==0.1.7
bidict==0.22.0
boto3==1.21.42
botocore==1.24.42
cachetools==5.0.0
certifi==2021.10.8
clip==1.0
colorama==0.4.4
cycler==0.11.0
cython==0.29.28
dnspython==2.1.0
docker-pycreds==0.4.0
dominate==2.6.0
easydict==1.9
einops==0.4.1
external==0.0.1
flags==0.0.1.2
fonttools==4.31.2
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gitdb==4.0.9
gitpython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.45.0
imageio==2.20.0
ipdb==0.13.9
ipykernel==6.2.0
ipywidgets==7.7.0
jmespath==1.0.0
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.3
jupyterlab-widgets==1.1.1
kiwisolver==1.4.2
lightning-utilities==0.6.0.post0
llvmlite==0.38.0
lmdb==1.3.0
markdown==3.3.6
matplotlib==3.5.1
multidict==6.0.4
networkx==2.6.3
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
onnx==1.11.0
onnxruntime-gpu==1.14.1
opencv-python==4.5.5.64
ordered-set==4.1.0
pandas==1.3.5
pathtools==0.1.2
phx-class-registry==3.0.5
pip==22.1.2
promise==2.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pylatex==1.4.1
pyparsing==3.0.7
pyqt5==5.12
pyqt5-sip==4.19.19
python-etcd==0.4.5
pytorch-lightning==1.9.0
pytorch-model-summary==0.1.2
pytorch-pretrained-bert==0.6.2
pytracking==0.2.2
pywavelets==1.3.0
pyzmq==23.2.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
scikit-image==0.19.3
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
smmap==5.0.0
some-package==0.1
tb-nightly==2.9.0a20220426
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.0.31-2005241907
tifffile==2021.11.2
tikzplotlib==0.10.1
timm==0.6.7
toml==0.10.2
torch==1.13.1
torchelastic==0.2.0
torchfile==0.1.0
torchmetrics==0.11.1
torchvision==0.14.1
transformers==4.18.0
urllib3==1.26.12
visdom==0.1.8.9
vit-pytorch==0.35.2
vot-toolkit==0.5.3
vot-trax==3.0.3
wandb==0.13.2
webcolors==1.11.1
werkzeug==2.1.1
widgetsnbextension==3.6.1
yacs==0.1.8
yarl==1.8.2
